requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth[with_social]
django-cors-headers
drf-spectacular
celery
redis
psycopg2-binary
django-redis